Replace unsafe use of QSharedPointer with reference
It's now the client's responsibility to keep the CorpusReader alive.
The old situation was unsafe because of the following.

Suppose the client code keeps a shared ptr to the reader. The FilterIter
then gets a "bare" ptr and makes a shared ptr, which keeps a separate
reference count from the original shared ptr (unless the debugging flag
QT_SHAREDPOINTER_TRACK_POINTERS is on).

When either of the reference counts goes to 0, the reader is destroyed,
leaving one of the sets of shared ptrs dangling.