capi: Change the API for alpinocorpus_iter_next.
Return 1 when the iterator moved and 0 when we have reached the end
of iteration. Not returning a pointer on each iteration makes it easier
to manage resources in the Haskell binding using ForeignPtr.

The Snap framework kills threads that run longer than the timeout,
so manual memory management was not safe there.

To consider: in some of the CString allocations, we still use manual
memory management. There is still a window for leaks when threads are
killed here.