Lazy directory reading. Fixes #2.
I also removed support for caching. Caching has been problematic ever since
we've done recursive reading of directories. Adding/removing a directory entry
only changes the timestamp of the parent directory, not of other ancestors.

To do proper caching, we have to check the timestamps on directories recursively,
which defies the point.