xpath: Add support for non-colored bracketing.
This is based on a patch provided by Peter Kleiweg. This changes the
behavior of alpinocorpus-xpath to print ASCII brackets ([]).  Colored
bracketing will only be used when the '-c' flag is specified.

I decided not to automatically enable coloring based on isatty(3),
because this function is not available on VC++. Moreover, it does not
work in pipelines where one would wish coloring (e.g. less -R).